requirements.txt
flask
python-dotenv
google-cloud-vision
openai
pillow